Split utils into those that need an atomspace,
and those that don't.  Otherwise, one gets into a circular
dependency linking hassle.
