+++ opencog/CMakeLists.txt
@@ -12,8 +12,9 @@ ADD_SUBDIRECTORY (learning)
 ADD_SUBDIRECTORY (reasoning)
 
 IF (HAVE_ATOMSPACE)
-    ADD_SUBDIRECTORY (atomspace)
     ADD_SUBDIRECTORY (atoms)
+    ADD_SUBDIRECTORY (atomspace)
+    ADD_SUBDIRECTORY (atomspaceutils)
     ADD_SUBDIRECTORY (atomutils)
     ADD_SUBDIRECTORY (benchmark)
     ADD_SUBDIRECTORY (query)
