+++ opencog/benchmark/module/CMakeLists.txt
@@ -6,5 +6,6 @@ ADD_DEPENDENCIES(benchmark opencog_atom_types)
 ADD_DEPENDENCIES(benchmark attention_atom_types)
 
 TARGET_LINK_LIBRARIES(benchmark
+	atomspaceutils
 	server
 )
