+++ opencog/dynamics/attention/HebbianCreationModule.cc
@@ -25,7 +25,7 @@
 
 #include <tbb/task.h>
 
-#include <opencog/atomutils/AtomSpaceUtils.h>
+#include <opencog/atomutils/AtomUtils.h>
 #include <opencog/atomspace/SimpleTruthValue.h>
 #include <opencog/dynamics/attention/atom_types.h>
 #include <opencog/server/CogServer.h>
